[knngraph] Static cast from size_t to int

+++ Ponca/src/SpatialPartitioning/KnnGraph/Query/knnGraphRangeQuery.h
@@ -46,7 +46,7 @@ public:
     }
 
     inline Iterator end(){
-        return Iterator(this, m_graph->size());
+        return Iterator(this, static_cast<int>(m_graph->size()));
     }
 
 protected:
